added videos_app logic inclued automatic creation of different formats (120p 240p etc)

+++ videos_app/signals.py
@@ -2,7 +2,7 @@ from .models import Video
 from django.dispatch import receiver
 from django.db.models.signals import post_save, post_delete
 import os
-from .tasks import convert_480p
+from .tasks import process_video_file
 import django_rq
 
 @receiver(post_save, sender=Video)
@@ -11,10 +11,9 @@ def video_post_save(sender, instance, created, **kwargs):
     if created:
         print(f"New video created: {instance.title}")
         queue = django_rq.get_queue('default', autocommit=True)
-        queue.enqueue(convert_480p, instance.video_file.path)
-    
+        queue.enqueue(process_video_file, instance.id)
+
 @receiver(post_delete, sender=Video)
 def video_post_delete(sender, instance, **kwargs):
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
-            os.remove(instance.video_file.path)
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        os.remove(instance.video_file.path)
